+++ src/wm/Client.cc
@@ -125,7 +125,11 @@ Client::Client(Window new_client, ClientInitConfig &initConfig, bool is_new)
 
 	// Construct the client
 	X11::grabServer();
-	if (! validate() || ! getAndUpdateWindowAttributes()) {
+	if (! validate()
+	    || ! getAndUpdateWindowAttributes()
+	    // Start of pekwm with already existing clients in WithdrawnState,
+	    // ignore these windows until they map
+	    || (pekwm::isStarting() && getWmState() == WithdrawnState)) {
 		X11::ungrabServer(true);
 		return;
 	}
@@ -1347,6 +1351,7 @@ Client::getWmState(void)
 	ulong items_read, items_left;
 	uchar *udata;
 
+	// 32bit state + 32bit icon window id
 	int status =
 		XGetWindowProperty(X11::getDpy(), _window,
 				   X11::getAtom(WM_STATE), 0L, 2L, False,